Separate the column renderers to make jest happy
