+++ src/automatic-annotations/shared/column-renderers/TaxonomicScope.tsx
@@ -1,5 +1,6 @@
 import { ExpandableList } from 'franklin-sites';
 
+import getLabelAndTooltip from '../../../shared/utils/getLabelAndTooltip';
 import TaxonomyView from '../../../shared/components/entry/TaxonomyView';
 
 import { ARBAAPIModel } from '../../arba/adapters/arbaConverter';
@@ -59,5 +60,15 @@ const TaxonomicScope = ({
     </ExpandableList>
   );
 };
+
+export function taxonomicScope() {
+  return {
+    ...getLabelAndTooltip(
+      'Taxonomic scope',
+      'The taxonomic scope to which the annotation rule applies'
+    ),
+    render: TaxonomicScope,
+  };
+}
 
 export default TaxonomicScope;
